Hide docs behind flag

+++ src/main.rs
@@ -58,6 +58,8 @@ enum Command {
     Build {
         #[structopt(help = "location of the project root", default_value = ".")]
         path: String,
+        #[structopt(help = "generate docs for this package as well", long)]
+        doc: bool,
     },
 
     #[structopt(name = "new", about = "Create a new project")]
@@ -108,7 +110,7 @@ struct ProjectConfig {
 
 fn main() {
     let result = match Command::from_args() {
-        Command::Build { path } => command_build(path),
+        Command::Build { path, doc } => command_build(path, doc),
 
         Command::Format {
             stdin,
@@ -130,7 +132,7 @@ fn main() {
     }
 }
 
-fn command_build(root: String) -> Result<(), Error> {
+fn command_build(root: String, write_docs: bool) -> Result<(), Error> {
     let mut srcs = vec![];
 
     // Read gleam.toml
@@ -164,8 +166,6 @@ fn command_build(root: String) -> Result<(), Error> {
 
     // Delete the gen directory before generating the newly compiled files
     let gen_dir = root_path.join("gen");
-    let doc_dir = root_path.join("doc");
-
     if gen_dir.exists() {
         std::fs::remove_dir_all(&gen_dir).map_err(|e| Error::FileIO {
             action: error::FileIOAction::Delete,
@@ -174,14 +174,16 @@ fn command_build(root: String) -> Result<(), Error> {
             err: Some(e.to_string()),
         })?;
     }
-
-    if doc_dir.exists() {
-        std::fs::remove_dir_all(&doc_dir.clone()).map_err(|e| Error::FileIO {
-            action: error::FileIOAction::Delete,
-            kind: error::FileKind::Directory,
-            path: doc_dir.clone(),
-            err: Some(e.to_string()),
-        })?;
+    let doc_dir = root_path.join("doc");
+    if write_docs {
+        if doc_dir.exists() {
+            std::fs::remove_dir_all(&doc_dir.clone()).map_err(|e| Error::FileIO {
+                action: error::FileIOAction::Delete,
+                kind: error::FileKind::Directory,
+                path: doc_dir.clone(),
+                err: Some(e.to_string()),
+            })?;
+        }
     }
 
     let mut doc_writer = doc::writer::DocWriter::new(
@@ -195,7 +197,9 @@ fn command_build(root: String) -> Result<(), Error> {
     );
 
     for crate::project::Compiled { files, doc, .. } in compiled {
-        doc_writer.add_chunk(doc);
+        if write_docs {
+            doc_writer.add_chunk(doc);
+        }
 
         for crate::project::OutputFile { text, path } in files {
             let dir_path = path.parent().ok_or_else(|| Error::FileIO {
@@ -229,9 +233,10 @@ fn command_build(root: String) -> Result<(), Error> {
     }
     println!("Done!");
 
-    doc_writer.write()?;
-
-    println!("Wrote docs to {}", doc_dir.to_str().unwrap());
+    if write_docs {
+        doc_writer.write()?;
+        println!("Wrote docs to {}", doc_dir.to_str().unwrap());
+    }
 
     Ok(())
 }
